refactor(kubernautagent): split investigate.go into focused files
Extracts handleStart+lease/upgrade helpers into investigate_start.go,
handleTakeover/handleMessage/handleComplete/handleStatus/handleReconnect
into investigate_takeover.go, handleDiscoverWorkflows+helpers into
investigate_discovery.go, and handleCancel/handleStartAutonomous+session
bookkeeping helpers into investigate_autonomous.go. Reduces investigate.go
from 1357 to 377 lines (types/interfaces/constructor/dispatch only), no
behavioral changes. gofmt realigned struct/const columns in
investigate_types.go and investigate_test.go incidentally.
