+++ internal/kubernautagent/mcp/tools/investigate.go
@@ -18,9 +18,6 @@ package tools
 
 import (
 	"context"
-	"encoding/json"
-	"errors"
-	"fmt"
 	"sync"
 	"time"
 
@@ -31,7 +28,6 @@ import (
 	"github.com/jordigilh/kubernaut/internal/kubernautagent/prompt"
 	"github.com/jordigilh/kubernaut/internal/kubernautagent/session"
 	katypes "github.com/jordigilh/kubernaut/pkg/kubernautagent/types"
-	"github.com/jordigilh/kubernaut/pkg/shared/transport"
 )
 
 // LLMMessage represents a single conversation message for the investigator.
@@ -378,980 +374,4 @@ func (t *InvestigateTool) dispatch(ctx context.Context, input InvestigateInput, 
 	default:
 		return InvestigateOutput{}, ErrInvalidAction
 	}
-}
-
-func (t *InvestigateTool) handleStart(ctx context.Context, input InvestigateInput, user mcpinternal.UserInfo) (InvestigateOutput, error) {
-	if t.rrChecker != nil {
-		exists, err := t.rrChecker.RemediationRequestExists(ctx, input.RRID)
-		if err != nil {
-			return InvestigateOutput{}, fmt.Errorf("validate remediation request: %w", err)
-		}
-		if !exists {
-			return InvestigateOutput{}, ErrCodeRRNotFound.WithDetail("rr_id", input.RRID)
-		}
-	}
-
-	// BR-INTERACTIVE-010: Check for pending interactive session and launch it.
-	// When launched, the investigation will self-transition to StatusUserDriving
-	// via InteractiveHold — skip TransitionToUserDriving below to avoid cancelling
-	// the RCA goroutine prematurely.
-	launchedPending, investigationSessionID := t.launchPendingInteractiveSession(input)
-
-	sess, startErr := t.startInteractiveSession(ctx, input, user)
-	if startErr != nil {
-		return InvestigateOutput{}, startErr
-	}
-
-	// #1390: Upgrade running autonomous session in-place (Jump In) instead of
-	// cancelling and recreating. UpgradeToInteractive sets the atomic flag so
-	// the goroutine's next InteractiveHold check sees it, and store.Update's
-	// deterministic check catches completion that already happened.
-	// Skip when we just launched a pending session — its RCA goroutine will
-	// self-transition via InteractiveHold once complete.
-	if !launchedPending {
-		investigationSessionID = t.upgradeOrCreateInteractiveSession(ctx, input, user)
-	}
-
-	if t.metrics != nil {
-		t.metrics.RecordInteractiveSessionStarted()
-		t.metrics.RecordInteractiveTakeover("start_success")
-	}
-
-	t.emitInteractiveStarted(sess.SessionID, input.RRID, user.Username)
-	t.startTimeoutTracking(sess.SessionID)
-	t.storeReconstructedContext(ctx, input.RRID, sess.SessionID)
-
-	return InvestigateOutput{
-		SessionID:              sess.SessionID,
-		Status:                 "started",
-		InvestigationSessionID: investigationSessionID,
-	}, nil
-}
-
-// launchPendingInteractiveSession launches a previously-deferred interactive
-// investigation, if one is pending for this session/RR. Returns whether a
-// pending session was launched and, when so, its session ID (used as the
-// InvestigationSessionID so the caller skips the autonomous-session upgrade
-// path below, since the deferred RCA goroutine self-transitions via
-// InteractiveHold once complete).
-func (t *InvestigateTool) launchPendingInteractiveSession(input InvestigateInput) (bool, string) {
-	pendingID := input.SessionID
-	hasPending := pendingID != ""
-	if !hasPending {
-		pendingID, hasPending = t.autoMgr.FindPendingByRemediationID(input.RRID)
-	}
-	if !hasPending {
-		return false, ""
-	}
-	if launchErr := t.autoMgr.LaunchDeferredInvestigation(pendingID); launchErr != nil {
-		t.logger.Error(launchErr, "start: failed to launch deferred investigation",
-			"rr_id", input.RRID, "pending_session_id", pendingID)
-		return false, ""
-	}
-	return true, pendingID
-}
-
-// acquireInteractiveLease takes over the interactive lease for rrID and maps
-// each failure mode (lease held, max sessions reached, or a generic takeover
-// error) to the appropriate MCP error and metric. raceLostMetric is recorded
-// when another driver holds the lease; failedMetric is recorded for every
-// other failure. genericErrPrefix labels the wrapped error text for the
-// generic-failure case (callers use distinct prefixes for start vs takeover).
-// Does not evaluate sess.Reconnected — callers handle the reconnect case
-// differently (handleStart rejects it, handleTakeover treats it as a
-// successful rejoin).
-func (t *InvestigateTool) acquireInteractiveLease(ctx context.Context, rrID string, user mcpinternal.UserInfo, raceLostMetric, failedMetric, genericErrPrefix string) (*mcpinternal.InteractiveSession, error) {
-	sess, err := t.sessions.Takeover(ctx, rrID, user)
-	if err == nil {
-		return sess, nil
-	}
-
-	if errors.Is(err, mcpinternal.ErrLeaseHeld) {
-		if t.metrics != nil {
-			t.metrics.RecordInteractiveLeaseContention()
-			t.metrics.RecordInteractiveTakeover(raceLostMetric)
-		}
-		driver, _ := t.sessions.GetDriver(rrID)
-		driverName := "unknown"
-		if driver != nil {
-			driverName = driver.ActingUser.Username
-		}
-		return nil, ErrCodeSessionActive.WithDetail("driver", driverName)
-	}
-	if errors.Is(err, mcpinternal.ErrMaxSessionsReached) {
-		if t.metrics != nil {
-			t.metrics.RecordInteractiveTakeover(failedMetric)
-		}
-		return nil, &MCPError{Code: "max_sessions", Message: "Maximum concurrent sessions reached"}
-	}
-	if t.metrics != nil {
-		t.metrics.RecordInteractiveTakeover(failedMetric)
-	}
-	return nil, fmt.Errorf("%s: %w", genericErrPrefix, err)
-}
-
-// startInteractiveSession takes over the interactive lease for the RR and
-// records the appropriate metrics/error mapping for each failure mode
-// (lease held, max sessions reached, reconnect-in-progress, or a generic
-// takeover error).
-func (t *InvestigateTool) startInteractiveSession(ctx context.Context, input InvestigateInput, user mcpinternal.UserInfo) (*mcpinternal.InteractiveSession, error) {
-	sess, err := t.acquireInteractiveLease(ctx, input.RRID, user, "start_failed", "start_failed", "start session")
-	if err != nil {
-		return nil, err
-	}
-
-	if sess.Reconnected {
-		return nil, &MCPError{
-			Code:    "session_active",
-			Message: "You already have an active session for this investigation; use action=reconnect to rejoin",
-			Details: map[string]string{
-				"driver":     user.Username,
-				"session_id": sess.SessionID,
-			},
-		}
-	}
-
-	return sess, nil
-}
-
-// upgradeOrCreateInteractiveSession upgrades the running autonomous session
-// for this RR in-place (Jump In), or — when no autonomous session exists, or
-// the existing one is terminal — creates a fresh interactive session so the
-// user is never left with a lease but no investigation to drive (#1440
-// SC-24). Returns the resulting investigation session ID.
-func (t *InvestigateTool) upgradeOrCreateInteractiveSession(ctx context.Context, input InvestigateInput, user mcpinternal.UserInfo) string {
-	autoSessionID, found := t.autoMgr.FindByRemediationID(input.RRID)
-	if !found {
-		// No session exists for this RR — create fresh interactive session so
-		// the user is never left with a lease but no investigation.
-		if freshID := t.createFallbackSession(ctx, input.RRID, user); freshID != "" {
-			return freshID
-		}
-		if forceErr := t.autoMgr.ForceTransitionToUserDriving(input.RRID, user.Username, user.Groups); forceErr != nil {
-			t.logger.Error(forceErr, "start: force-transition to user-driving (no running session found)",
-				"rr_id", input.RRID)
-		}
-		return ""
-	}
-
-	upgradeErr := t.autoMgr.UpgradeToInteractive(autoSessionID, user.Username, user.Groups)
-	if upgradeErr == nil {
-		return autoSessionID
-	}
-	if !errors.Is(upgradeErr, session.ErrSessionTerminal) {
-		t.logger.Error(upgradeErr, "start: upgrade autonomous session to interactive",
-			"rr_id", input.RRID, "auto_session_id", autoSessionID)
-		return autoSessionID
-	}
-
-	if forceErr := t.autoMgr.ForceTransitionToUserDriving(input.RRID, user.Username, user.Groups); forceErr != nil {
-		t.logger.Error(forceErr, "start: force-transition to user-driving failed (session terminal)",
-			"rr_id", input.RRID, "auto_session_id", autoSessionID)
-	}
-	// #1440 SC-24: Terminal session — create fresh interactive session so the
-	// user always has an investigation to drive.
-	if freshID := t.createFallbackSession(ctx, input.RRID, user); freshID != "" {
-		return freshID
-	}
-	return autoSessionID
-}
-
-// transitionAutonomousToUserDriving transitions the running autonomous
-// session for rrID to user-driven (#774: TransitionToUserDriving replaces
-// SuspendInvestigation so the session enters StatusUserDriving — pollable,
-// not terminal — with identity written to session metadata for the AA poll
-// response's Rego input.identity). When no autonomous session is found by RR
-// ID, retries ForceTransitionToUserDriving briefly to allow for the race
-// between MCP takeover and AA reconcile. Returns a non-nil error only for a
-// non-terminal TransitionToUserDriving failure; terminal-session and
-// not-found cases are logged and treated as best-effort.
-func (t *InvestigateTool) transitionAutonomousToUserDriving(rrID string, user mcpinternal.UserInfo) error {
-	autoSessionID, found := t.autoMgr.FindByRemediationID(rrID)
-	if !found {
-		// No running session found by RR ID. The AA session submit may still
-		// be in-flight (race between MCP takeover and AA reconcile). Retry
-		// briefly to allow the session to appear before giving up.
-		var forceErr error
-		for attempt := 0; attempt < 5; attempt++ {
-			forceErr = t.autoMgr.ForceTransitionToUserDriving(rrID, user.Username, user.Groups)
-			if forceErr == nil || !errors.Is(forceErr, session.ErrSessionNotFound) {
-				break
-			}
-			time.Sleep(100 * time.Millisecond)
-		}
-		if forceErr != nil {
-			t.logger.Error(forceErr, "takeover: force-transition to user-driving failed after retries",
-				"rr_id", rrID)
-		}
-		return nil
-	}
-
-	err := t.autoMgr.TransitionToUserDriving(autoSessionID, user.Username, user.Groups)
-	if err == nil {
-		return nil
-	}
-	if !errors.Is(err, session.ErrSessionTerminal) {
-		return fmt.Errorf("transition autonomous session to user-driving: %w", err)
-	}
-	if forceErr := t.autoMgr.ForceTransitionToUserDriving(rrID, user.Username, user.Groups); forceErr != nil {
-		t.logger.Error(forceErr, "takeover: force-transition to user-driving failed",
-			"rr_id", rrID, "auto_session_id", autoSessionID)
-	}
-	return nil
-}
-
-func (t *InvestigateTool) handleTakeover(ctx context.Context, input InvestigateInput, user mcpinternal.UserInfo) (InvestigateOutput, error) {
-	mu := t.getSessionMutex(input.RRID)
-	mu.Lock()
-	defer mu.Unlock()
-
-	// H4: Acquire the interactive Lease BEFORE suspending autonomous. This ensures
-	// that if Takeover fails (lease contention, max sessions), the autonomous
-	// investigation is NOT irreversibly cancelled.
-	sess, err := t.acquireInteractiveLease(ctx, input.RRID, user, "takeover_race_lost", "takeover_failed", "takeover session")
-	if err != nil {
-		return InvestigateOutput{}, err
-	}
-
-	if sess.Reconnected {
-		if t.timeoutTracker != nil {
-			t.timeoutTracker.ResetInactivity(sess.SessionID)
-		}
-		return InvestigateOutput{
-			SessionID: sess.SessionID,
-			Status:    "reconnected",
-		}, nil
-	}
-
-	// Lease acquired — now safe to transition autonomous investigation to user-driven.
-	if transitionErr := t.transitionAutonomousToUserDriving(input.RRID, user); transitionErr != nil {
-		if t.metrics != nil {
-			t.metrics.RecordInteractiveTakeover("takeover_failed")
-		}
-		return InvestigateOutput{}, transitionErr
-	}
-
-	if t.metrics != nil {
-		t.metrics.RecordInteractiveSessionStarted()
-		t.metrics.RecordInteractiveTakeover("takeover_success")
-	}
-
-	t.emitInteractiveStarted(sess.SessionID, input.RRID, user.Username)
-	t.startTimeoutTracking(sess.SessionID)
-
-	reconCount := t.storeReconstructedContext(ctx, input.RRID, sess.SessionID)
-	contextSummary := fmt.Sprintf("%d prior turns reconstructed", reconCount)
-
-	return InvestigateOutput{
-		SessionID: sess.SessionID,
-		Status:    "takeover_started",
-		Response:  contextSummary,
-	}, nil
-}
-
-func (t *InvestigateTool) handleMessage(ctx context.Context, input InvestigateInput, user mcpinternal.UserInfo) (InvestigateOutput, error) {
-	mu := t.getSessionMutex(input.RRID)
-	mu.Lock()
-	defer mu.Unlock()
-
-	sess, authErr := t.authorizeActiveDriver(input.RRID, user)
-	if authErr != nil {
-		return InvestigateOutput{}, authErr
-	}
-
-	// SEC-HIGH-01: Enforce per-session message rate limit before processing.
-	if t.rateLimiter != nil {
-		if rlErr := t.rateLimiter.Allow(sess.SessionID, len(input.Message)); rlErr != nil {
-			if errors.Is(rlErr, mcpinternal.ErrRateLimited) {
-				return InvestigateOutput{}, ErrCodeRateLimited
-			}
-			return InvestigateOutput{}, ErrCodeRateLimited
-		}
-	}
-
-	// SEC-04: Touch activity to reset inactivity timer.
-	t.sessions.TouchActivity(input.RRID)
-	if t.timeoutTracker != nil {
-		t.timeoutTracker.ResetInactivity(sess.SessionID)
-	}
-
-	// #898-S5: Attach session ID for audit attribution on K8s API calls.
-	ctx = transport.WithAuditSessionID(ctx, sess.SessionID)
-
-	// F9 / #1374: Attach signal context for PhaseRCA tool parity with
-	// the autonomous path. Future tools may read SignalContextFromContext.
-	if t.signalResolver != nil {
-		if resolved, resolveErr := t.signalResolver.ResolveSignalContext(ctx, input.RRID); resolveErr == nil && resolved != nil {
-			ctx = katypes.WithSignalContext(ctx, *resolved)
-		}
-	}
-
-	// Clear DiscoveryResult before the LLM call: any message after
-	// discover_workflows invalidates stale recommendations, forcing re-discovery
-	// before select_workflow can be called.
-	if sess.DiscoveryResult != nil {
-		sess.DiscoveryResult = nil
-	}
-
-	// PROD-02: Prepend reconstructed context turns so the LLM has full history.
-	messages := t.buildMessagesWithContext(input.RRID, input.Message)
-
-	response, err := t.runner.RunInteractiveTurn(ctx, messages, input.RRID)
-	if err != nil {
-		return InvestigateOutput{}, fmt.Errorf("interactive turn failed: %w", err)
-	}
-
-	// Accumulate the user message + LLM response in reconHistory so that
-	// subsequent actions (discover_workflows) can extract RCA from the
-	// full conversation without relying on audit trace reconstruction.
-	t.appendConversationTurn(input.RRID, input.Message, response)
-
-	// Reset inactivity timer AFTER the LLM call completes. The pre-call reset
-	// (above) prevents timeout during user think-time; this post-call reset
-	// prevents timeout during slow LLM responses that exceed the inactivity
-	// window. Without this, a 90s LLM response with a 60s timeout would
-	// expire the session mid-turn.
-	t.sessions.TouchActivity(input.RRID)
-	if t.timeoutTracker != nil {
-		t.timeoutTracker.ResetInactivity(sess.SessionID)
-	}
-
-	return InvestigateOutput{
-		SessionID: sess.SessionID,
-		Status:    "message_received",
-		Response:  response,
-	}, nil
-}
-
-func (t *InvestigateTool) handleComplete(input InvestigateInput, user mcpinternal.UserInfo) (InvestigateOutput, error) {
-	mu := t.getSessionMutex(input.RRID)
-	mu.Lock()
-	defer mu.Unlock()
-
-	if !t.sessions.IsDriverActive(input.RRID) {
-		return InvestigateOutput{}, ErrCodeNotFound
-	}
-
-	sess, err := t.sessions.GetDriver(input.RRID)
-	if err != nil || sess == nil {
-		return InvestigateOutput{}, ErrCodeNotFound
-	}
-
-	// SEC-CRIT-01: Only the active driver may terminate the session.
-	if sess.ActingUser.Username != user.Username {
-		return InvestigateOutput{}, ErrCodeSessionActive.WithDetail("driver", sess.ActingUser.Username)
-	}
-
-	if t.timeoutTracker != nil {
-		t.timeoutTracker.StopTracking(sess.SessionID)
-	}
-
-	if err := t.sessions.Release(sess.SessionID, "complete"); err != nil {
-		if errors.Is(err, mcpinternal.ErrSessionNotFound) {
-			// H3: Session already released (race with timeout/disconnect), but the
-			// user sees "completed" — still emit audit for completeness.
-			t.emitInteractiveCompleted(sess.SessionID, input.RRID, user.Username, "complete_already_released")
-			return InvestigateOutput{SessionID: sess.SessionID, Status: "completed"}, nil
-		}
-		return InvestigateOutput{}, fmt.Errorf("release session: %w", err)
-	}
-
-	t.emitInteractiveCompleted(sess.SessionID, input.RRID, user.Username, "complete")
-
-	var finalResult *katypes.InvestigationResult
-	if sess.RCAResult != nil {
-		r := *sess.RCAResult
-		r.Reason = "investigation completed by user"
-		finalResult = &r
-	} else {
-		finalResult = &katypes.InvestigationResult{
-			RCASummary: "Investigation completed without workflow selection",
-			Reason:     "investigation completed by user",
-		}
-	}
-	notActionable := false
-	finalResult.IsActionable = &notActionable
-	finalResult.Warnings = append(finalResult.Warnings, "Alert not actionable")
-
-	CompleteHTTPSession(t.httpCompleter, input.RRID, finalResult, t.logger, "complete")
-
-	if t.metrics != nil {
-		t.metrics.RecordInteractiveSessionEnded()
-	}
-
-	t.sessionMu.Delete(input.RRID)
-	t.reconHistory.Delete(input.RRID)
-
-	return InvestigateOutput{
-		SessionID: sess.SessionID,
-		Status:    "completed",
-	}, nil
-}
-
-func (t *InvestigateTool) handleStatus(input InvestigateInput, user mcpinternal.UserInfo) (InvestigateOutput, error) {
-	status := StatusOutput{RRID: input.RRID}
-
-	if t.sessions.IsDriverActive(input.RRID) {
-		driver, _ := t.sessions.GetDriver(input.RRID)
-		status.Mode = StatusModeInteractive
-		if driver != nil && driver.ActingUser.Username == user.Username {
-			status.Driver = driver.ActingUser.Username
-		}
-	} else if _, found := t.autoMgr.FindByRemediationID(input.RRID); found {
-		status.Mode = StatusModeAutonomous
-	} else {
-		status.Mode = StatusModeNotFound
-	}
-
-	data, err := json.Marshal(status)
-	if err != nil {
-		return InvestigateOutput{}, fmt.Errorf("marshal status: %w", err)
-	}
-
-	return InvestigateOutput{
-		Status:   "status",
-		Response: string(data),
-	}, nil
-}
-
-func (t *InvestigateTool) handleReconnect(input InvestigateInput, user mcpinternal.UserInfo) (InvestigateOutput, error) {
-	sess, authErr := t.authorizeActiveDriver(input.RRID, user)
-	if authErr != nil {
-		return InvestigateOutput{}, authErr
-	}
-
-	t.sessions.TouchActivity(input.RRID)
-	if t.timeoutTracker != nil {
-		t.timeoutTracker.ResetInactivity(sess.SessionID)
-	}
-
-	return InvestigateOutput{
-		SessionID: sess.SessionID,
-		Status:    "reconnected",
-	}, nil
-}
-
-func (t *InvestigateTool) handleDiscoverWorkflows(ctx context.Context, input InvestigateInput, user mcpinternal.UserInfo) (InvestigateOutput, error) {
-	mu := t.getSessionMutex(input.RRID)
-	mu.Lock()
-	defer mu.Unlock()
-
-	sess, authErr := t.authorizeActiveDriver(input.RRID, user)
-	if authErr != nil {
-		return InvestigateOutput{}, authErr
-	}
-	if t.catalog == nil {
-		return InvestigateOutput{}, fmt.Errorf("workflow catalog not configured: cannot enrich discovery names")
-	}
-
-	// Reset inactivity timer before the (potentially long) LLM calls.
-	t.sessions.TouchActivity(input.RRID)
-	if t.timeoutTracker != nil {
-		t.timeoutTracker.ResetInactivity(sess.SessionID)
-	}
-
-	// Step 1: Obtain the structured RCA result for Phase 3 workflow discovery.
-	rcaResult, err := t.resolveRCAForDiscovery(ctx, input.RRID, sess.SessionID)
-	if err != nil {
-		return InvestigateOutput{}, err
-	}
-
-	// Step 2: Resolve signal context for Phase 3. Enrichment is handled
-	// internally by the investigator's enrichment pipeline (F5 #1374), so we
-	// only resolve the signal here.
-	signal := t.resolveDiscoverySignal(ctx, input.RRID)
-
-	// Step 3: Enrich context with the HTTP investigation session so that
-	// workflow discovery can emit audit events with session_id and stream
-	// events to the subscriber (#1384: Bug A fix).
-	ctx = t.enrichDiscoveryContext(ctx, input.RRID)
-
-	// Step 4: Run Phase 3 workflow discovery using the structured RCA.
-	// The investigator resolves enrichment internally when its enricher is wired,
-	// falling back to nil when not available.
-	workflowResult, err := t.runner.RunWorkflowDiscovery(ctx, signal, rcaResult, nil, input.RRID)
-	if err != nil {
-		return InvestigateOutput{}, fmt.Errorf("workflow discovery failed: %w", err)
-	}
-
-	// Step 5: Store results on the interactive session.
-	sess.RCAResult = rcaResult
-	sess.DiscoveryResult = extractDiscoveryResult(workflowResult)
-
-	// Step 6: Populate discovery target visibility fields (#1437).
-	populateDiscoveryTargets(sess.DiscoveryResult, signal, workflowResult.RemediationTarget)
-
-	// Enrich workflow names from catalog.
-	t.enrichDiscoveryNames(ctx, sess.DiscoveryResult)
-
-	// Reset inactivity timer after the LLM calls complete.
-	t.sessions.TouchActivity(input.RRID)
-	if t.timeoutTracker != nil {
-		t.timeoutTracker.ResetInactivity(sess.SessionID)
-	}
-
-	// Build the JSON response for the user.
-	discoveryJSON, err := json.Marshal(sess.DiscoveryResult)
-	if err != nil {
-		return InvestigateOutput{}, fmt.Errorf("marshal discovery result: %w", err)
-	}
-
-	return InvestigateOutput{
-		SessionID: sess.SessionID,
-		Status:    "workflows_discovered",
-		Response:  string(discoveryJSON),
-	}, nil
-}
-
-// authorizeActiveDriver verifies the requesting user is the active driver of
-// the interactive session for rrID, returning ErrCodeNotDriving /
-// ErrCodeSessionExpired / ErrCodeSessionActive as appropriate. Shared by the
-// message/reconnect/discover_workflows actions, which all reject non-drivers
-// identically.
-func (t *InvestigateTool) authorizeActiveDriver(rrID string, user mcpinternal.UserInfo) (*mcpinternal.InteractiveSession, error) {
-	if !t.sessions.IsDriverActive(rrID) {
-		return nil, ErrCodeNotDriving
-	}
-
-	sess, err := t.sessions.GetDriver(rrID)
-	if err != nil {
-		if errors.Is(err, mcpinternal.ErrSessionExpired) {
-			return nil, ErrCodeSessionExpired
-		}
-		return nil, ErrCodeNotDriving
-	}
-	if sess == nil {
-		return nil, ErrCodeNotDriving
-	}
-
-	if sess.ActingUser.Username != user.Username {
-		return nil, ErrCodeSessionActive.WithDetail("driver", sess.ActingUser.Username)
-	}
-
-	return sess, nil
-}
-
-// resolveRCAForDiscovery obtains the structured RCA result used for Phase 3
-// workflow discovery.
-//
-// Preferred path: reuse the full InvestigationResult already produced by the
-// autonomous Phase 1 RCA and stored in the session manager. This preserves
-// the complete RemediationTarget (Kind, APIVersion, Name, Namespace) that
-// the LLM emitted during the original investigation.
-//
-// Fallback: if no stored result exists (e.g. pure interactive session),
-// reconstruct conversation from audit traces and re-extract RCA.
-func (t *InvestigateTool) resolveRCAForDiscovery(ctx context.Context, rrID, sessionID string) (*katypes.InvestigationResult, error) {
-	if storedResult, ok := t.autoMgr.GetLatestRCAResultByRemediationID(rrID); ok && storedResult != nil {
-		t.logger.Info("discover_workflows: using stored RCA result from autonomous investigation",
-			"rr_id", rrID,
-			"rca_target_kind", storedResult.RemediationTarget.Kind,
-			"rca_target_api_version", storedResult.RemediationTarget.APIVersion,
-			"rca_target_name", storedResult.RemediationTarget.Name)
-		return storedResult, nil
-	}
-
-	messages := t.buildMessagesWithContext(rrID, "")
-	if len(messages) > 0 && messages[len(messages)-1].Content == "" {
-		messages = messages[:len(messages)-1]
-	}
-	if len(messages) == 0 {
-		reconCount := t.storeReconstructedContext(ctx, rrID, sessionID)
-		t.logger.Info("discover_workflows: reconHistory was empty, reconstructed from audit traces",
-			"rr_id", rrID, "recon_turns", reconCount)
-		messages = t.buildMessagesWithContext(rrID, "")
-		if len(messages) > 0 && messages[len(messages)-1].Content == "" {
-			messages = messages[:len(messages)-1]
-		}
-	}
-
-	if len(messages) == 0 {
-		t.logger.Info("discover_workflows: no conversation context available after reconstruction",
-			"rr_id", rrID)
-		return nil, fmt.Errorf("rca extraction failed: no conversation context available — investigation audit traces not found in data storage")
-	}
-
-	rcaResult, err := t.runner.RunRCAExtraction(ctx, messages, rrID)
-	if err != nil {
-		return nil, fmt.Errorf("rca extraction failed: %w", err)
-	}
-
-	// Phase 2 extraction from conversation reconstructs a best-effort RCA,
-	// but its RemediationTarget is unreliable: the conversation messages lack
-	// the system prompt (with signal name/resource), so the LLM may fall back
-	// to a generic target. Clear it so RunWorkflowDiscoveryFromRCA preserves
-	// the signal resolver's authoritative identity instead of overwriting it
-	// via SyncSignalFromRCA with the extraction's guess.
-	rcaResult.RemediationTarget = katypes.RemediationTarget{}
-	return rcaResult, nil
-}
-
-// resolveDiscoverySignal resolves the signal context for Phase 3 workflow
-// discovery, logging (but not failing) on resolution errors.
-func (t *InvestigateTool) resolveDiscoverySignal(ctx context.Context, rrID string) katypes.SignalContext {
-	var signal katypes.SignalContext
-	if t.signalResolver == nil {
-		return signal
-	}
-	resolved, resolveErr := t.signalResolver.ResolveSignalContext(ctx, rrID)
-	if resolveErr != nil {
-		t.logger.V(1).Info("signal context resolution failed, using empty context",
-			"rr_id", rrID, "error", resolveErr)
-	} else if resolved != nil {
-		signal = *resolved
-	}
-	return signal
-}
-
-// enrichDiscoveryContext attaches the HTTP investigation session ID (and its
-// lazy audit sink, when available) to ctx so workflow discovery emits audit
-// events with session_id and streams events to the subscriber (#1384: Bug A
-// fix).
-func (t *InvestigateTool) enrichDiscoveryContext(ctx context.Context, rrID string) context.Context {
-	if t.httpCompleter == nil {
-		return ctx
-	}
-	httpSessionID, found := t.httpCompleter.FindUserDrivingByRemediationID(rrID)
-	if !found {
-		return ctx
-	}
-	ctx = session.WithSessionID(ctx, httpSessionID)
-	if ls, ok := t.autoMgr.GetSessionLazySink(httpSessionID); ok {
-		ctx = session.WithLazySink(ctx, ls)
-	}
-	t.logger.V(1).Info("discover_workflows: enriched context with HTTP session",
-		"rr_id", rrID, "http_session_id", httpSessionID)
-	return ctx
-}
-
-// populateDiscoveryTargets fills in the SignalTarget/SearchedTarget
-// visibility fields on dr (#1437). SignalTarget is always the original alert
-// resource. SearchedTarget is the resource actually searched against the
-// catalog — sourced from the RCA-resolved remediation target, falling back
-// to the signal target when empty (e.g. fallback RCA path).
-func populateDiscoveryTargets(dr *mcpinternal.WorkflowDiscoveryResult, signal katypes.SignalContext, rt katypes.RemediationTarget) {
-	signalTarget := &mcpinternal.DiscoveryTargetInfo{
-		APIVersion: signal.ResourceAPIVersion,
-		Kind:       signal.ResourceKind,
-		Name:       signal.ResourceName,
-		Namespace:  signal.Namespace,
-	}
-	dr.SignalTarget = signalTarget
-
-	if rt.Kind == "" {
-		dr.SearchedTarget = signalTarget
-		return
-	}
-	dr.SearchedTarget = &mcpinternal.DiscoveryTargetInfo{
-		APIVersion: rt.APIVersion,
-		Kind:       rt.Kind,
-		Name:       rt.Name,
-		Namespace:  rt.Namespace,
-	}
-}
-
-// extractDiscoveryResult builds a WorkflowDiscoveryResult from the Phase 3
-// InvestigationResult, separating the selected workflow from alternatives.
-func extractDiscoveryResult(result *katypes.InvestigationResult) *mcpinternal.WorkflowDiscoveryResult {
-	if result == nil {
-		return &mcpinternal.WorkflowDiscoveryResult{}
-	}
-
-	dr := &mcpinternal.WorkflowDiscoveryResult{
-		FullResult: result,
-	}
-
-	if result.WorkflowID != "" {
-		dr.Recommended = &mcpinternal.DiscoveredWorkflow{
-			WorkflowID:      result.WorkflowID,
-			ExecutionBundle: result.ExecutionBundle,
-			Confidence:      result.Confidence,
-			Rationale:       result.WorkflowRationale,
-			Parameters:      cloneParameterMap(result.Parameters),
-		}
-	}
-
-	if len(result.AlternativeWorkflows) > 0 {
-		dr.Alternatives = make([]mcpinternal.DiscoveredWorkflow, 0, len(result.AlternativeWorkflows))
-		for _, alt := range result.AlternativeWorkflows {
-			dr.Alternatives = append(dr.Alternatives, mcpinternal.DiscoveredWorkflow{
-				WorkflowID:      alt.WorkflowID,
-				ExecutionBundle: alt.ExecutionBundle,
-				Confidence:      alt.Confidence,
-				Rationale:       alt.Rationale,
-				Parameters:      cloneParameterMap(alt.Parameters),
-			})
-		}
-	}
-
-	return dr
-}
-
-// enrichDiscoveryNames resolves human-readable workflow names from the catalog
-// for each discovered workflow. Lookup failures are logged but do not fail the
-// operation (the workflow is still usable by ID).
-func (t *InvestigateTool) enrichDiscoveryNames(ctx context.Context, dr *mcpinternal.WorkflowDiscoveryResult) {
-	if dr == nil {
-		return
-	}
-	if dr.Recommended != nil {
-		dr.Recommended.Name = t.resolveWorkflowName(ctx, dr.Recommended.WorkflowID)
-	}
-	for i := range dr.Alternatives {
-		dr.Alternatives[i].Name = t.resolveWorkflowName(ctx, dr.Alternatives[i].WorkflowID)
-	}
-}
-
-func (t *InvestigateTool) resolveWorkflowName(ctx context.Context, workflowID string) string {
-	if workflowID == "" {
-		return ""
-	}
-	wf, err := t.catalog.GetWorkflowByID(ctx, workflowID)
-	if err != nil {
-		t.logger.V(1).Info("catalog lookup failed for workflow name enrichment",
-			"workflow_id", workflowID, "error", err)
-		return ""
-	}
-	return wf.WorkflowName
-}
-
-func (t *InvestigateTool) handleCancel(input InvestigateInput, user mcpinternal.UserInfo) (InvestigateOutput, error) {
-	mu := t.getSessionMutex(input.RRID)
-	mu.Lock()
-	defer mu.Unlock()
-
-	if !t.sessions.IsDriverActive(input.RRID) {
-		return InvestigateOutput{}, ErrNoActiveSession
-	}
-
-	sess, err := t.sessions.GetDriver(input.RRID)
-	if err != nil || sess == nil {
-		return InvestigateOutput{}, ErrNoActiveSession
-	}
-
-	// SEC-CRIT-01: Only the active driver may cancel the session.
-	if sess.ActingUser.Username != user.Username {
-		return InvestigateOutput{}, ErrCodeSessionActive.WithDetail("driver", sess.ActingUser.Username)
-	}
-
-	if t.timeoutTracker != nil {
-		t.timeoutTracker.StopTracking(sess.SessionID)
-	}
-
-	if err := t.sessions.Release(sess.SessionID, "explicit"); err != nil {
-		if !errors.Is(err, mcpinternal.ErrSessionNotFound) {
-			return InvestigateOutput{}, fmt.Errorf("release session: %w", err)
-		}
-	}
-
-	t.emitInteractiveCompleted(sess.SessionID, input.RRID, user.Username, "cancel")
-
-	CompleteHTTPSession(t.httpCompleter, input.RRID, nil, t.logger, "cancel")
-
-	if t.metrics != nil {
-		t.metrics.RecordInteractiveSessionEnded()
-	}
-
-	t.sessionMu.Delete(input.RRID)
-	t.reconHistory.Delete(input.RRID)
-
-	return InvestigateOutput{
-		SessionID: sess.SessionID,
-		Status:    "cancelled",
-	}, nil
-}
-
-func (t *InvestigateTool) handleStartAutonomous(ctx context.Context, input InvestigateInput, _ mcpinternal.UserInfo) (InvestigateOutput, error) {
-	if t.rrChecker != nil {
-		exists, err := t.rrChecker.RemediationRequestExists(ctx, input.RRID)
-		if err != nil {
-			return InvestigateOutput{}, fmt.Errorf("validate remediation request: %w", err)
-		}
-		if !exists {
-			return InvestigateOutput{}, ErrCodeRRNotFound.WithDetail("rr_id", input.RRID)
-		}
-	}
-
-	if existingID, found := t.autoMgr.FindByRemediationID(input.RRID); found {
-		return InvestigateOutput{
-			SessionID: existingID,
-			Status:    "already_running",
-		}, nil
-	}
-
-	metadata := map[string]string{
-		"remediation_id": input.RRID,
-	}
-
-	// F4 (#1374): Resolve signal context to build a real InvestigateFunc.
-	// The autonomous investigation uses the same full pipeline as the HTTP path.
-	if t.signalResolver == nil {
-		return InvestigateOutput{}, fmt.Errorf("start autonomous investigation: signal resolver not configured")
-	}
-	resolved, resolveErr := t.signalResolver.ResolveSignalContext(ctx, input.RRID)
-	if resolveErr != nil {
-		return InvestigateOutput{}, fmt.Errorf("resolve signal context for autonomous investigation: %w", resolveErr)
-	}
-	if resolved == nil {
-		return InvestigateOutput{}, fmt.Errorf("start autonomous investigation: no signal context for rr_id %s", input.RRID)
-	}
-	signal := *resolved
-
-	sessionID, err := t.autoMgr.StartInvestigation(ctx, func(bgCtx context.Context) (*katypes.InvestigationResult, error) {
-		return t.runner.RunFullInvestigation(bgCtx, signal)
-	}, metadata)
-	if err != nil {
-		if errors.Is(err, session.ErrMaxInvestigationsReached) {
-			return InvestigateOutput{}, ErrCodeMaxInvestigations
-		}
-		return InvestigateOutput{}, fmt.Errorf("start autonomous investigation: %w", err)
-	}
-
-	if _, subErr := t.autoMgr.Subscribe(ctx, sessionID); subErr != nil {
-		t.logger.Error(subErr, "start_autonomous: Subscribe failed, events may be lost",
-			"session_id", sessionID, "rr_id", input.RRID)
-	}
-
-	return InvestigateOutput{
-		SessionID: sessionID,
-		Status:    "autonomous_started",
-	}, nil
-}
-
-// createFallbackSession creates a fresh interactive session when no viable
-// autonomous session exists (no session found, or terminal session). This
-// ensures the user always has an investigation to drive after acquiring the
-// MCP lease (SC-24, #1440).
-func (t *InvestigateTool) createFallbackSession(ctx context.Context, rrID string, user mcpinternal.UserInfo) string {
-	metadata := map[string]string{
-		"rr_id":    rrID,
-		"username": user.Username,
-		"mode":     "interactive_fallback",
-	}
-	investigateFn := session.InvestigateFunc(func(_ context.Context) (*katypes.InvestigationResult, error) {
-		return &katypes.InvestigationResult{
-			RCASummary:      "Interactive session — awaiting user direction",
-			InteractiveHold: true,
-		}, nil
-	})
-	sessionID, err := t.autoMgr.StartInvestigation(ctx, investigateFn, metadata)
-	if err != nil {
-		t.logger.Error(err, "start: fallback session creation failed",
-			"rr_id", rrID, "username", user.Username)
-		return ""
-	}
-	return sessionID
-}
-
-// storeReconstructedContext queries the reconstructor and caches prior turns
-// for the session's lifetime. Prefers RCA summary from a prior completed
-// session (more concise, prevents token bloat) over full audit trail
-// reconstruction. Returns the number of turns stored.
-func (t *InvestigateTool) storeReconstructedContext(ctx context.Context, rrID, sessionID string) int {
-	// BR-INTERACTIVE-010: If a prior session produced an RCA summary, use it
-	// as a concise seed instead of reconstructing the full audit trail.
-	if rcaSummary, hasRCA := t.autoMgr.GetLatestRCASummaryByRemediationID(rrID); hasRCA {
-		history := []LLMMessage{
-			{Role: "assistant", Content: "Previous investigation RCA summary: " + rcaSummary},
-		}
-		t.reconHistory.Store(rrID, history)
-		return 1
-	}
-
-	turns, reconErr := t.recon.Reconstruct(ctx, rrID, sessionID)
-	if reconErr != nil {
-		t.logger.Error(reconErr, "context reconstruction from DS failed, proceeding with empty context",
-			"rr_id", rrID, "session_id", sessionID)
-	}
-	if len(turns) == 0 {
-		return 0
-	}
-
-	history := make([]LLMMessage, 0, len(turns))
-	for _, turn := range turns {
-		if turn.Content == "" {
-			continue
-		}
-		history = append(history, LLMMessage{Role: turn.Role, Content: turn.Content})
-	}
-	if len(history) == 0 {
-		return 0
-	}
-	t.reconHistory.Store(rrID, history)
-	return len(history)
-}
-
-// appendConversationTurn appends a user message and the LLM response to
-// reconHistory so that discover_workflows can extract RCA from the
-// accumulated interactive conversation.
-func (t *InvestigateTool) appendConversationTurn(rrID, userMessage, assistantResponse string) {
-	var history []LLMMessage
-	if raw, ok := t.reconHistory.Load(rrID); ok {
-		history = raw.([]LLMMessage)
-	}
-	history = append(history,
-		LLMMessage{Role: "user", Content: userMessage},
-		LLMMessage{Role: "assistant", Content: assistantResponse},
-	)
-	t.reconHistory.Store(rrID, history)
-}
-
-// buildMessagesWithContext prepends any cached reconstruction history to the
-// current user message, giving the LLM full prior context (PROD-02).
-// Copies the cached slice to avoid aliasing the sync.Map entry.
-func (t *InvestigateTool) buildMessagesWithContext(rrID, userMessage string) []LLMMessage {
-	var history []LLMMessage
-	if raw, ok := t.reconHistory.Load(rrID); ok {
-		cached := raw.([]LLMMessage)
-		history = make([]LLMMessage, len(cached))
-		copy(history, cached)
-	}
-	messages := make([]LLMMessage, 0, len(history)+1)
-	messages = append(messages, history...)
-	messages = append(messages, LLMMessage{Role: "user", Content: userMessage})
-	return messages
-}
-
-// emitInteractiveStarted emits aiagent.interactive.started (BR-INTERACTIVE-003, DD-INTERACTIVE-002).
-// Uses context.Background() because audit is fire-and-forget (ADR-038) and must not
-// be tied to the request lifecycle — a cancelled request context must not drop the event.
-func (t *InvestigateTool) emitInteractiveStarted(sessionID, correlationID, actingUser string) {
-	if t.auditStore == nil {
-		return
-	}
-	event := audit.NewEvent(audit.EventTypeInteractiveStarted, correlationID,
-		audit.WithSessionID(sessionID),
-		audit.WithActingUser(actingUser),
-	)
-	event.EventAction = audit.ActionInteractiveStarted
-	event.EventOutcome = audit.OutcomeSuccess
-	audit.StoreBestEffort(context.Background(), t.auditStore, event, t.logger)
-}
-
-// emitInteractiveCompleted emits aiagent.interactive.completed (BR-INTERACTIVE-003, DD-INTERACTIVE-002).
-func (t *InvestigateTool) emitInteractiveCompleted(sessionID, correlationID, actingUser, reason string) {
-	if t.auditStore == nil {
-		return
-	}
-	event := audit.NewEvent(audit.EventTypeInteractiveCompleted, correlationID,
-		audit.WithSessionID(sessionID),
-		audit.WithActingUser(actingUser),
-	)
-	event.EventAction = audit.ActionInteractiveCompleted
-	event.EventOutcome = audit.OutcomeSuccess
-	event.Data["reason"] = reason
-	audit.StoreBestEffort(context.Background(), t.auditStore, event, t.logger)
-}
-
-// startTimeoutTracking begins inactivity tracking for a session if configured.
-func (t *InvestigateTool) startTimeoutTracking(sessionID string) {
-	if t.timeoutTracker == nil {
-		return
-	}
-	notify := func(msg string) {
-		if t.notifyFn != nil {
-			t.notifyFn(sessionID, msg)
-		}
-	}
-	t.timeoutTracker.StartTracking(sessionID, notify)
 }
