+++ internal/kubernautagent/mcp/tools/investigate_types.go
@@ -21,7 +21,7 @@ import "errors"
 // InvestigateInput is the JSON schema for the kubernaut_investigate MCP tool.
 type InvestigateInput struct {
 	RRID             string   `json:"rr_id"`
-	Action           string   `json:"action"` // start, message, complete, cancel, takeover, status, reconnect, discover_workflows
+	Action           string   `json:"action"`               // start, message, complete, cancel, takeover, status, reconnect, discover_workflows
 	SessionID        string   `json:"session_id,omitempty"` // #1452: AF-provided KA session ID for direct lookup
 	Message          string   `json:"message,omitempty"`
 	ActingUser       string   `json:"acting_user,omitempty"`
@@ -52,15 +52,15 @@ type StatusOutput struct {
 
 // Valid actions for the kubernaut_investigate tool.
 const (
-	ActionStart              = "start"
-	ActionMessage            = "message"
-	ActionComplete           = "complete"
-	ActionCancel             = "cancel"
-	ActionTakeover           = "takeover"
-	ActionStatus             = "status"
-	ActionReconnect          = "reconnect"
-	ActionDiscoverWorkflows  = "discover_workflows"
-	ActionStartAutonomous    = "start_autonomous"
+	ActionStart             = "start"
+	ActionMessage           = "message"
+	ActionComplete          = "complete"
+	ActionCancel            = "cancel"
+	ActionTakeover          = "takeover"
+	ActionStatus            = "status"
+	ActionReconnect         = "reconnect"
+	ActionDiscoverWorkflows = "discover_workflows"
+	ActionStartAutonomous   = "start_autonomous"
 )
 
 var (
